Setup application components.
Adjust operation result conversion.
